--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -1,12 +1,14 @@
 package hw10programoptimization
 
 import (
-	"encoding/json"
+	"bufio"
+	"bytes"
 	"fmt"
 	"io"
-	"io/ioutil"
-	"regexp"
 	"strings"
+	"sync"
+
+	"github.com/goccy/go-json"
 )
 
 type User struct {
@@ -21,47 +23,57 @@ type User struct {
 
 type DomainStat map[string]int
 
+var userPool = sync.Pool{
+	New: func() interface{} {
+		return &User{}
+	},
+}
+
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
-	u, err := getUsers(r)
+	u, err := getUsers(r, domain)
 	if err != nil {
 		return nil, fmt.Errorf("get users error: %w", err)
 	}
-	return countDomains(u, domain)
+
+	return countDomains(u)
 }
 
-type users [100_000]User
+type users []User
 
-func getUsers(r io.Reader) (result users, err error) {
-	content, err := ioutil.ReadAll(r)
-	if err != nil {
-		return
-	}
+func getUsers(r io.Reader, domain string) (result users, err error) {
+	scanner := bufio.NewScanner(r)
 
-	lines := strings.Split(string(content), "\n")
-	for i, line := range lines {
-		var user User
-		if err = json.Unmarshal([]byte(line), &user); err != nil {
+	for scanner.Scan() {
+		if !bytes.Contains(scanner.Bytes(), []byte("."+domain)) {
+			continue
+		}
+
+		user := *userPool.Get().(*User)
+		userPool.Put(&user)
+
+		if err = json.Unmarshal(scanner.Bytes(), &user); err != nil {
 			return
 		}
-		result[i] = user
+
+		result = append(result, user)
 	}
+
 	return
 }
 
-func countDomains(u users, domain string) (DomainStat, error) {
-	result := make(DomainStat)
+func countDomains(u users) (DomainStat, error) {
+	result := make(DomainStat, len(u))
 
 	for _, user := range u {
-		matched, err := regexp.Match("\\."+domain, []byte(user.Email))
-		if err != nil {
-			return nil, err
+		at := strings.LastIndex(user.Email, "@")
+
+		if at == 0 {
+			continue
 		}
 
-		if matched {
-			num := result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]
-			num++
-			result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])] = num
-		}
+		nextDomainLvl := strings.ToLower(user.Email[at+1:])
+
+		result[nextDomainLvl]++
 	}
 	return result, nil
 }